# backend/app/api/static/docs/openapi-3.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "Track, Manage, and Organize your Things.",
        "title": "Homebox API",
        "contact": {
            "name": "Homebox Team",
            "url": "https://discord.homebox.software"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/v1/users": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Get All Users",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/repo.UserAdminOut"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Create User",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/services.UserAdminCreate"
                            }
                        }
                    },
                    "description": "User Data",
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/repo.UserAdminOut"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/users/change-password": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "User"
                ],
                "summary": "Change Password",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/v1.ChangePassword"
                            }
                        }
                    },
                    "description": "Password Payload",
                    "required": true
                },
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
        "/v1/users/register": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "First-Time Setup: Create Administrator Account",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/services.UserRegistration"
                            }
                        }
                    },
                    "description": "User Data",
                    "required": true
                },
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "403": {
                        "description": "Setup already completed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/v1/users/self": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get User Self",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/v1.Wrapped"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "item": {
                                                    "$ref": "#/components/schemas/v1.UserSelfOut"
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "User"
                ],
                "summary": "Update Account",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/repo.UserUpdate"
                            }
                        }
                    },
                    "description": "User Data",
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/v1.Wrapped"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "item": {
                                                    "$ref": "#/components/schemas/repo.UserUpdate"
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "User"
                ],
                "summary": "Delete Account",
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
        "/v1/users/self/settings": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get user settings",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/v1.Wrapped"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "item": {
                                                    "type": "object",
                                                    "additionalProperties": true
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "User"
                ],
                "summary": "Update user settings",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    },
                    "description": "Settings Data",
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/v1.Wrapped"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "item": {
                                                    "type": "object",
                                                    "additionalProperties": true
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/users/{id}": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Update User (name, email, roles, optional password reset)",
                "parameters": [
                    {
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/services.UserAdminUpdate"
                            }
                        }
                    },
                    "description": "User Data",
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/repo.UserAdminOut"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Delete User",
                "parameters": [
                    {
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        }
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "components": {
        "securitySchemes": {
            "Bearer": {
                "description": "\"Type 'Bearer TOKEN' to correctly set the API Key\"",
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        },
        "schemas": {
...
            "repo.RoleSummary": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "isSuperAdmin": {
                        "type": "boolean"
                    },
                    "name": {
                        "type": "string"
                    }
                }
            },
...
            "repo.UserAdminOut": {
                "type": "object",
                "properties": {
                    "defaultGroupId": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "oidcIssuer": {
                        "type": "string"
                    },
                    "oidcSubject": {
                        "type": "string"
                    },
                    "roles": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/repo.RoleSummary"
                        }
                    }
                }
            },
            "repo.UserUpdate": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    }
                }
            },
...
            "services.UserAdminCreate": {
                "type": "object",
                "required": [
                    "email",
                    "name",
                    "password"
                ],
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "password": {
                        "type": "string"
                    },
                    "roleIds": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
            "services.UserAdminUpdate": {
                "type": "object",
                "required": [
                    "email",
                    "name"
                ],
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "password": {
                        "description": "Password optionally resets the user's password when non-empty.",
                        "type": "string"
                    },
                    "roleIds": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
            "services.UserRegistration": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                }
            },
...
            "v1.ChangePassword": {
                "type": "object",
                "properties": {
                    "current": {
                        "type": "string"
                    },
                    "new": {
                        "type": "string"
                    }
                }
            },
...
            "v1.PermissionGrantOut": {
                "type": "object",
                "properties": {
                    "collectionId": {
                        "type": "string",
                        "nullable": true
                    },
                    "create": {
                        "type": "boolean"
                    },
                    "delete": {
                        "type": "boolean"
                    },
                    "edit": {
                        "type": "boolean"
                    },
                    "section": {
                        "type": "string"
                    },
                    "view": {
                        "type": "boolean"
                    }
                }
            },
...
            "v1.UserSelfOut": {
                "type": "object",
                "properties": {
                    "defaultGroupId": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "isSuperAdmin": {
                        "type": "boolean"
                    },
                    "name": {
                        "type": "string"
                    },
                    "oidcIssuer": {
                        "type": "string"
                    },
                    "oidcSubject": {
                        "type": "string"
                    },
                    "permissions": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/v1.PermissionGrantOut"
                        }
                    },
                    "roles": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/repo.RoleSummary"
                        }
                    }
                }
            },
            "v1.Wrapped": {
                "type": "object",
                "properties": {
                    "item": {}
                }
            },
...
        }
    }
}